refactor(cmd): thin list.go; move logic to internal/listing
- New internal/listing package owns runs/workspaces/pipelines/personas
  /adapters/contracts/skills/compositions listing domain logic, filter,
  sort, and JSON output schemas.
- cmd/wave/commands/list.go reduced from 1630 LOC to 324 LOC; the new
  list_render.go owns table rendering. CLI behaviour, flag names,
  command names, and JSON output schemas are unchanged.
- skills.go now imports internal/listing for CollectSkillPipelineUsage.

Partial of #1502 (list.go portion). init.go/run.go/compose.go thin
deferred to follow-up PRs.
