+++ internal/listing/duration.go
@@ -1,0 +1,26 @@
+package listing
+
+import (
+	"fmt"
+	"time"
+)
+
+// FormatDuration renders a duration as a compact human-readable string:
+// sub-second values use milliseconds, sub-minute values use fractional seconds,
+// sub-hour values use mNNs, and longer values use hHHmMM.
+func FormatDuration(d time.Duration) string {
+	if d < time.Second {
+		return fmt.Sprintf("%dms", d.Milliseconds())
+	}
+	if d < time.Minute {
+		return fmt.Sprintf("%.1fs", d.Seconds())
+	}
+	if d < time.Hour {
+		mins := int(d.Minutes())
+		secs := int(d.Seconds()) % 60
+		return fmt.Sprintf("%dm%ds", mins, secs)
+	}
+	hours := int(d.Hours())
+	mins := int(d.Minutes()) % 60
+	return fmt.Sprintf("%dh%dm", hours, mins)
+}
